requirements.txt
prometheus_client
kafka
elasticsearch
python-dotenv
pytz